Struct google_metastore1_beta::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the DataprocMetastore hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_metastore1_beta as metastore1_beta;
 
use std::default::Default;
use oauth2;
use metastore1_beta::DataprocMetastore;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = DataprocMetastore::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `locations_get(...)`, `locations_list(...)`, `locations_operations_delete(...)`, `locations_operations_get(...)`, `locations_operations_list(...)`, `locations_services_backups_create(...)`, `locations_services_backups_delete(...)`, `locations_services_backups_get(...)`, `locations_services_backups_list(...)`, `locations_services_create(...)`, `locations_services_delete(...)`, `locations_services_export_metadata(...)`, `locations_services_get(...)`, `locations_services_get_iam_policy(...)`, `locations_services_list(...)`, `locations_services_metadata_imports_create(...)`, `locations_services_metadata_imports_get(...)`, `locations_services_metadata_imports_list(...)`, `locations_services_metadata_imports_patch(...)`, `locations_services_patch(...)`, `locations_services_restore(...)`, `locations_services_set_iam_policy(...)` and `locations_services_test_iam_permissions(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn locations_operations_delete(
&self,
name: &str
) -> ProjectLocationOperationDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

Arguments

name - The name of the operation resource to be deleted.

pub fn locations_operations_get(
&self,
name: &str
) -> ProjectLocationOperationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn locations_operations_list(
&self,
name: &str
) -> ProjectLocationOperationListCall<'a, C>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED.NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users//operations. To override the binding, API services can add a binding such as "/v1/{name=users/}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Arguments

name - The name of the operation's parent resource.

pub fn locations_services_backups_create(
&self,
request: Backup,
parent: &str
) -> ProjectLocationServiceBackupCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new Backup in a given project and location.

Arguments

request - No description provided.

parent - Required. The relative resource name of the service in which to create a backup of the following form:projects/{project_number}/locations/{location_id}/services/{service_id}.

pub fn locations_services_backups_delete(
&self,
name: &str
) -> ProjectLocationServiceBackupDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a single backup.

Arguments

name - Required. The relative resource name of the backup to delete, in the following form:projects/{project_number}/locations/{location_id}/services/{service_id}/backups/{backup_id}.

pub fn locations_services_backups_get(
&self,
name: &str
) -> ProjectLocationServiceBackupGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of a single backup.

Arguments

name - Required. The relative resource name of the backup to retrieve, in the following form:projects/{project_number}/locations/{location_id}/services/{service_id}/backups/{backup_id}.

pub fn locations_services_backups_list(
&self,
parent: &str
) -> ProjectLocationServiceBackupListCall<'a, C>

Create a builder to help you perform the following task:

Lists backups in a service.

Arguments

parent - Required. The relative resource name of the service whose backups to list, in the following form:projects/{project_number}/locations/{location_id}/services/{service_id}/backups.

pub fn locations_services_metadata_imports_create(
&self,
request: MetadataImport,
parent: &str
) -> ProjectLocationServiceMetadataImportCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new MetadataImport in a given project and location.

Arguments

request - No description provided.

parent - Required. The relative resource name of the service in which to create a metastore import, in the following form:projects/{project_number}/locations/{location_id}/services/{service_id}.

pub fn locations_services_metadata_imports_get(
&self,
name: &str
) -> ProjectLocationServiceMetadataImportGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of a single import.

Arguments

name - Required. The relative resource name of the metadata import to retrieve, in the following form:projects/{project_number}/locations/{location_id}/services/{service_id}/metadataImports/{import_id}.

pub fn locations_services_metadata_imports_list(
&self,
parent: &str
) -> ProjectLocationServiceMetadataImportListCall<'a, C>

Create a builder to help you perform the following task:

Lists imports in a service.

Arguments

parent - Required. The relative resource name of the service whose metadata imports to list, in the following form:projects/{project_number}/locations/{location_id}/services/{service_id}/metadataImports.

pub fn locations_services_metadata_imports_patch(
&self,
request: MetadataImport,
name: &str
) -> ProjectLocationServiceMetadataImportPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a single import. Only the description field of MetadataImport is supported to be updated.

Arguments

request - No description provided.

name - Immutable. The relative resource name of the metadata import, of the form:projects/{project_number}/locations/{location_id}/services/{service_id}/metadataImports/{metadata_import_id}.

pub fn locations_services_create(
&self,
request: Service,
parent: &str
) -> ProjectLocationServiceCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a metastore service in a project and location.

Arguments

request - No description provided.

parent - Required. The relative resource name of the location in which to create a metastore service, in the following form:projects/{project_number}/locations/{location_id}.

pub fn locations_services_delete(
&self,
name: &str
) -> ProjectLocationServiceDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a single service.

Arguments

name - Required. The relative resource name of the metastore service to delete, in the following form:projects/{project_number}/locations/{location_id}/services/{service_id}.

pub fn locations_services_export_metadata(
&self,
request: ExportMetadataRequest,
service: &str
) -> ProjectLocationServiceExportMetadataCall<'a, C>

Create a builder to help you perform the following task:

Exports metadata from a service.

Arguments

request - No description provided.

service - Required. The relative resource name of the metastore service to run export, in the following form:projects/{project_id}/locations/{location_id}/services/{service_id}.

pub fn locations_services_get(
&self,
name: &str
) -> ProjectLocationServiceGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the details of a single service.

Arguments

name - Required. The relative resource name of the metastore service to retrieve, in the following form:projects/{project_number}/locations/{location_id}/services/{service_id}.

pub fn locations_services_get_iam_policy(
&self,
resource: &str
) -> ProjectLocationServiceGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_services_list(
&self,
parent: &str
) -> ProjectLocationServiceListCall<'a, C>

Create a builder to help you perform the following task:

Lists services in a project and location.

Arguments

parent - Required. The relative resource name of the location of metastore services to list, in the following form:projects/{project_number}/locations/{location_id}.

pub fn locations_services_patch(
&self,
request: Service,
name: &str
) -> ProjectLocationServicePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the parameters of a single service.

Arguments

request - No description provided.

name - Immutable. The relative resource name of the metastore service, of the form:projects/{project_number}/locations/{location_id}/services/{service_id}.

pub fn locations_services_restore(
&self,
request: RestoreServiceRequest,
service: &str
) -> ProjectLocationServiceRestoreCall<'a, C>

Create a builder to help you perform the following task:

Restores a service from a backup.

Arguments

request - No description provided.

service - Required. The relative resource name of the metastore service to run restore, in the following form:projects/{project_id}/locations/{location_id}/services/{service_id}.

pub fn locations_services_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationServiceSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy.Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_services_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationServiceTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error.Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_get(&self, name: &str) -> ProjectLocationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about a location.

Arguments

name - Resource name for the location.

pub fn locations_list(&self, name: &str) -> ProjectLocationListCall<'a, C>

Create a builder to help you perform the following task:

Lists information about the supported locations for this service.

Arguments

name - The resource that owns the locations collection, if applicable.